# src/api/swagger/models/release_extended.go
package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	strfmt "github.com/go-openapi/strfmt"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/validate"
)

/*ReleaseExtended release extended

swagger:model releaseExtended
*/
type ReleaseExtended struct {

	/* chart name

	Required: true
	*/
	ChartName *string `json:"chartName"`

	/* chart version

	Required: true
	*/
	ChartVersion *string `json:"chartVersion"`

	/* name

	Required: true
	*/
	Name *string `json:"name"`

	/* namespace

	Required: true
	*/
	Namespace *string `json:"namespace"`

	/* notes

	Required: true
	*/
	Notes *string `json:"notes"`

	/* resources

	Required: true
	*/
	Resources *string `json:"resources"`

	/* status

	Required: true
	*/
	Status *string `json:"status"`

	/* updated

	Required: true
	*/
	Updated *string `json:"updated"`
}

// Validate validates this release extended
func (m *ReleaseExtended) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateChartName(formats); err != nil {
		// prop
		res = append(res, err)
	}

	if err := m.validateChartVersion(formats); err != nil {
		// prop
		res = append(res, err)
	}

	if err := m.validateName(formats); err != nil {
		// prop
		res = append(res, err)
	}

	if err := m.validateNamespace(formats); err != nil {
		// prop
		res = append(res, err)
	}

	if err := m.validateNotes(formats); err != nil {
		// prop
		res = append(res, err)
	}

	if err := m.validateResources(formats); err != nil {
		// prop
		res = append(res, err)
	}

	if err := m.validateStatus(formats); err != nil {
		// prop
		res = append(res, err)
	}

	if err := m.validateUpdated(formats); err != nil {
		// prop
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *ReleaseExtended) validateChartName(formats strfmt.Registry) error {

	if err := validate.Required("chartName", "body", m.ChartName); err != nil {
		return err
	}

	return nil
}

func (m *ReleaseExtended) validateChartVersion(formats strfmt.Registry) error {

	if err := validate.Required("chartVersion", "body", m.ChartVersion); err != nil {
		return err
	}

	return nil
}

func (m *ReleaseExtended) validateName(formats strfmt.Registry) error {

	if err := validate.Required("name", "body", m.Name); err != nil {
		return err
	}

	return nil
}

func (m *ReleaseExtended) validateNamespace(formats strfmt.Registry) error {

	if err := validate.Required("namespace", "body", m.Namespace); err != nil {
		return err
	}

	return nil
}

func (m *ReleaseExtended) validateNotes(formats strfmt.Registry) error {

	if err := validate.Required("notes", "body", m.Notes); err != nil {
		return err
	}

	return nil
}

func (m *ReleaseExtended) validateResources(formats strfmt.Registry) error {

	if err := validate.Required("resources", "body", m.Resources); err != nil {
		return err
	}

	return nil
}

func (m *ReleaseExtended) validateStatus(formats strfmt.Registry) error {

	if err := validate.Required("status", "body", m.Status); err != nil {
		return err
	}

	return nil
}

func (m *ReleaseExtended) validateUpdated(formats strfmt.Registry) error {

	if err := validate.Required("updated", "body", m.Updated); err != nil {
		return err
	}

	return nil
}